added hook checkTimeDifference
added hook checkTimeDifference + refactored code using this hook
